LayoutTests/fast/js/dfg-call-method-hit-watchpoint-expected.txt

Tests correctness of method calls when the prototype is changed.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS callFoo(o) is 42
PASS callFoo(o) is 42
PASS callFoo(o) is 42
PASS callFoo(o) is 42
PASS callFoo(o) is 42
PASS callFoo(o) is 42
PASS callFoo(o) is 42
PASS callFoo(o) is 42
PASS callFoo(o) is 42
PASS callFoo(o) is 42
PASS callFoo(o) is 42
PASS callFoo(o) is 42
PASS callFoo(o) is 42
PASS callFoo(o) is 42
PASS callFoo(o) is 42
PASS callFoo(o) is 42
PASS callFoo(o) is 42
PASS callFoo(o) is 42
PASS callFoo(o) is 42
PASS callFoo(o) is 42
PASS callFoo(o) is 42
PASS callFoo(o) is 42
PASS callFoo(o) is 42
PASS callFoo(o) is 42
PASS callFoo(o) is 42
PASS callFoo(o) is 42
PASS callFoo(o) is 42
PASS callFoo(o) is 42
PASS callFoo(o) is 42
PASS callFoo(o) is 42
PASS callFoo(o) is 42
PASS callFoo(o) is 42
PASS callFoo(o) is 42
PASS callFoo(o) is 42
PASS callFoo(o) is 42
PASS callFoo(o) is 42
PASS callFoo(o) is 42
PASS callFoo(o) is 42
PASS callFoo(o) is 42
PASS callFoo(o) is 42
PASS callFoo(o) is 42
PASS callFoo(o) is 42
PASS callFoo(o) is 42
PASS callFoo(o) is 42
PASS callFoo(o) is 42
PASS callFoo(o) is 42
PASS callFoo(o) is 42
PASS callFoo(o) is 42
PASS callFoo(o) is 42
PASS callFoo(o) is 42
PASS callFoo(o) is 42
PASS callFoo(o) is 42
PASS callFoo(o) is 42
PASS callFoo(o) is 42
PASS callFoo(o) is 42
PASS callFoo(o) is 42
PASS callFoo(o) is 42
PASS callFoo(o) is 42
PASS callFoo(o) is 42
PASS callFoo(o) is 42
PASS callFoo(o) is 42
PASS callFoo(o) is 42
PASS callFoo(o) is 42
PASS callFoo(o) is 42
PASS callFoo(o) is 42
PASS callFoo(o) is 42
PASS callFoo(o) is 42
PASS callFoo(o) is 42
PASS callFoo(o) is 42
PASS callFoo(o) is 42
PASS callFoo(o) is 42
PASS callFoo(o) is 42
PASS callFoo(o) is 42
PASS callFoo(o) is 42
PASS callFoo(o) is 42
PASS callFoo(o) is 42
PASS callFoo(o) is 42
PASS callFoo(o) is 42
PASS callFoo(o) is 42
PASS callFoo(o) is 42
PASS callFoo(o) is 42
PASS callFoo(o) is 42
PASS callFoo(o) is 42
PASS callFoo(o) is 42
PASS callFoo(o) is 42
PASS callFoo(o) is 42
PASS callFoo(o) is 42
PASS callFoo(o) is 42
PASS callFoo(o) is 42
PASS callFoo(o) is 42
PASS callFoo(o) is 42
PASS callFoo(o) is 42
PASS callFoo(o) is 42
PASS callFoo(o) is 42
PASS callFoo(o) is 42
PASS callFoo(o) is 42
PASS callFoo(o) is 42
PASS callFoo(o) is 42
PASS callFoo(o) is 42
PASS callFoo(o) is 42
PASS callFoo(o) is 42
PASS callFoo(o) is 42
PASS callFoo(o) is 42
PASS callFoo(o) is 42
PASS callFoo(o) is 42
PASS callFoo(o) is 42
PASS callFoo(o) is 42
PASS callFoo(o) is 42
PASS callFoo(o) is 42
PASS callFoo(o) is 42
PASS callFoo(o) is 42
PASS callFoo(o) is 42
PASS callFoo(o) is 42
PASS callFoo(o) is 42
PASS callFoo(o) is 42
PASS callFoo(o) is 42
PASS callFoo(o) is 42
PASS callFoo(o) is 42
PASS callFoo(o) is 42
PASS callFoo(o) is 42
PASS callFoo(o) is 42
PASS callFoo(o) is 42
PASS callFoo(o) is 42
PASS callFoo(o) is 42
PASS callFoo(o) is 42
PASS callFoo(o) is 42
PASS callFoo(o) is 42
PASS callFoo(o) is 42
PASS callFoo(o) is 42
PASS callFoo(o) is 42
PASS callFoo(o) is 42
PASS callFoo(o) is 42
PASS callFoo(o) is 42
PASS callFoo(o) is 42
PASS callFoo(o) is 42
PASS callFoo(o) is 42
PASS callFoo(o) is 42
PASS callFoo(o) is 42
PASS callFoo(o) is 42
PASS callFoo(o) is 42
PASS callFoo(o) is 42
PASS callFoo(o) is 42
PASS callFoo(o) is 42
PASS callFoo(o) is 42
PASS callFoo(o) is 42
PASS callFoo(o) is 42
PASS callFoo(o) is 42
PASS callFoo(o) is 42
PASS callFoo(o) is 42
PASS callFoo(o) is 42
PASS callFoo(o) is 56
PASS callFoo(o) is 56
PASS callFoo(o) is 56
PASS callFoo(o) is 56
PASS callFoo(o) is 56
PASS callFoo(o) is 56
PASS callFoo(o) is 56
PASS callFoo(o) is 56
PASS callFoo(o) is 56
PASS callFoo(o) is 56
PASS callFoo(o) is 56
PASS callFoo(o) is 56
PASS callFoo(o) is 56
PASS callFoo(o) is 56
PASS callFoo(o) is 56
PASS callFoo(o) is 56
PASS callFoo(o) is 56
PASS callFoo(o) is 56
PASS callFoo(o) is 56
PASS callFoo(o) is 56
PASS callFoo(o) is 56
PASS callFoo(o) is 56
PASS callFoo(o) is 56
PASS callFoo(o) is 56
PASS callFoo(o) is 56
PASS callFoo(o) is 56
PASS callFoo(o) is 56
PASS callFoo(o) is 56
PASS callFoo(o) is 56
PASS callFoo(o) is 56
PASS callFoo(o) is 56
PASS callFoo(o) is 56
PASS callFoo(o) is 56
PASS callFoo(o) is 56
PASS callFoo(o) is 56
PASS callFoo(o) is 56
PASS callFoo(o) is 56
PASS callFoo(o) is 56
PASS callFoo(o) is 56
PASS callFoo(o) is 56
PASS callFoo(o) is 56
PASS callFoo(o) is 56
PASS callFoo(o) is 56
PASS callFoo(o) is 56
PASS callFoo(o) is 56
PASS callFoo(o) is 56
PASS callFoo(o) is 56
PASS callFoo(o) is 56
PASS callFoo(o) is 56
PASS callFoo(o) is 56
PASS successfullyParsed is true

TEST COMPLETE
